Moved dev preview deeper into tae server init
