+++ packages/app/src/cli/services/dev/processes/types.ts
@@ -1,4 +1,5 @@
-import {AbortSignal} from '@shopify/cli-kit/node/abort'
+import type {AbortSignal} from '@shopify/cli-kit/node/abort'
+import type {AdminSession} from '@shopify/cli-kit/node/session'
 
 import {Writable} from 'node:stream'
 
@@ -11,4 +12,15 @@ export interface BaseProcess<T> {
   prefix: string
   function: DevProcessFunction<T>
   options: T
+}
+
+export interface PreviewThemeAppExtensionsOptions {
+  adminSession: AdminSession
+  themeExtensionServerArgs: string[]
+  storefrontToken: string
+  token: string
+}
+
+export interface PreviewThemeAppExtensionsProcess extends BaseProcess<PreviewThemeAppExtensionsOptions> {
+  type: 'theme-app-extensions'
 }
